Improve performance of finding observations
Instead of searching through all observations in one big array and then 
finding the one with matching ID, we store observations as an object. 
The observation IDs are the object keys and the observations are the 
corresponding values. The performance of object[key] is much better than 
array.find(e => e.id === key). See https://stackoverflow.com/a/17295727 
and https://jsben.ch/Y9jDP for details.