Timeseries: Encode values as array
Instead of storing timeseries data for a specified period (e.g.,
2000-2021), we store timeseries an array where the first element is the
first year with data, the second element is the value for the first
year, and all subsequent elements are the values for the subsequent
years (e.g., third element is the value for the year after the first
year, etc...).

This allows timeseries to handle an arbitrary time period and in most
cases greatly reduces the data size.