Fix isDataLoaded: Return false if data is a pending promise
Only return true for isDataLoaded if data is a plain object or an array.
Promises will return false.

When we start loading data, we store the pending promise on the data
object, so that future requests to load the data can reuse the pending
promise (rather than triggering another load). Therefore, our
isDataLoaded check needs to make sure that data is actually a plain
object or an array before returning true.